backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
kafka-python==2.0.2
pydantic~=2.10.3
SQLAlchemy~=2.0.36
python-decouple~=3.8
passlib~=1.7.4
jose~=1.0.0
email_validator~=2.2.0
python-dotenv~=1.0.1